Move built-in generators to Generators folder
